redo whitelist command as per Raymond suggestion
